refactor: delegate application API discovery to a dedicated component
